+++ Lmyc/Models/Volunteer.cs
@@ -23,13 +23,13 @@ namespace Lmyc.Models
         [DisplayFormat(DataFormatString = "{0:MM/dd/yyyy}", ApplyFormatInEditMode = true)]
         public DateTime Date { get; set; }
 
-        [Required(ErrorMessage = "Duration in hours is requried")]
-        [Range(1, 24, ErrorMessage = "Duration of voluntter work have to be between 1 to 24 hours")]
+        [Required(ErrorMessage = "Duration in hours is required")]
+        [Range(1, 24, ErrorMessage = "Duration of volunteer work have to be between 1 to 24 hours")]
         [DisplayName("Number of Hours")]
         public int Duration { get; set; }
 
-        [Required(ErrorMessage = "Description of the Voluntter work is required")]
-        [Range(20, 2048, ErrorMessage = "Description of voluntter work have to be between 20 to 2048 letters")]
+        [Required(ErrorMessage = "Description of the Volunteer work is required")]
+        [MaxLength(2048, ErrorMessage = "Description of volunteer work have to be between 20 to 2048 letters")]
         public string Description { get; set; }
 
         [Required(ErrorMessage = "Please select at least one classification code")]